Rename HousingContainer to Housing
